refactor(plan-lifecycle): migrate ralphai.ts + subcommands to plan-lifecycle facade
