+++ src/seed.ts
@@ -1,7 +1,7 @@
 import { existsSync, mkdirSync, writeFileSync, readFileSync, rmSync } from "fs";
 import { join } from "path";
 import { RESET, DIM, TEXT } from "./utils.ts";
-import { getRepoPipelineDirs } from "./global-state.ts";
+import { getRepoPipelineDirs } from "./plan-lifecycle.ts";
 import { getConfigFilePath } from "./config.ts";
 
 // ---------------------------------------------------------------------------
